+++ library/GitHub/Workflow/Command/Syntax/Key.hs
@@ -1,4 +1,4 @@
-module GitHub.Workflow.Command.Properties.Key
+module GitHub.Workflow.Command.Syntax.Key
   ( Key
   ) where
 
@@ -7,8 +7,8 @@ import Control.Lens (iso)
 import Data.String (IsString)
 import Data.Text (Text)
 import Data.Text.Encoding qualified as T
-import GitHub.Workflow.Command.TextIso
-import GitHub.Workflow.Command.ToByteStringBuilder
+import GitHub.Workflow.Command.Syntax.TextIso
+import GitHub.Workflow.Command.Syntax.ToByteStringBuilder
 import Prelude (Eq, Ord, Show)
 
 newtype Key = Key {text :: Text}
